+++ be/src/routes/Transaction_VND.route.ts
@@ -1,17 +1,26 @@
 import express from 'express';
-import { PrismaClient } from '@prisma/client';
+import { PrismaClient, Role } from '@prisma/client';
+import { verifyToken, AuthRequest } from '../middleware/VerifyToken';
 
 const router = express.Router();
 const prisma = new PrismaClient();
 
 // Get all VND transactions
-router.get('/', async (req, res) => {
+router.get('/', verifyToken, async (req: AuthRequest, res) => {
+    if (!req.user) {
+        res.status(401).json({ error: 'Unauthorized: Please log in first' });
+        return;
+    }
     const transactions = await prisma.transaction_VND.findMany();
     res.json(transactions);
 });
 
 // Get VND transaction by ID
-router.get('/:id', async (req, res) => {
+router.get('/:id', verifyToken, async (req: AuthRequest, res) => {
+    if (!req.user) {
+        res.status(401).json({ error: 'Unauthorized: Please log in first' });
+        return;
+    }
     const transaction = await prisma.transaction_VND.findUnique({
         where: { id: req.params.id }
     });
@@ -19,7 +28,11 @@ router.get('/:id', async (req, res) => {
 });
 
 // Create VND transaction
-router.post('/', async (req, res) => {
+router.post('/', verifyToken, async (req: AuthRequest, res) => {
+    if (!req.user) {
+        res.status(401).json({ error: 'Unauthorized: Please log in first' });
+        return;
+    }
     const { buyer_id, model_id, amount_vnd, proof_image, status } = req.body;
 
     try {
@@ -34,7 +47,12 @@ router.post('/', async (req, res) => {
 });
 
 // Update VND transaction
-router.put('/:id', async (req, res) => {
+router.put('/:id', verifyToken, async (req: AuthRequest, res) => {
+    const { user_id, role } = req.user!;
+    if (req.user?.role !== 'admin') {
+        res.status(401).json({ error: 'access denied' });
+        return;
+    }
     const { amount_vnd, proof_image, status } = req.body;
 
     try {
@@ -50,9 +68,14 @@ router.put('/:id', async (req, res) => {
 });
 
 // Delete VND transaction
-router.delete('/:id', async (req, res) => {
+router.delete('/:id', verifyToken, async (req: AuthRequest, res) => {
+    console.log('Delete VND transaction');
     const transactionId = req.params.id;
-
+    const { user_id, role } = req.user!;
+    if (req.user?.role !== 'admin') {
+        res.status(401).json({ error: 'access denied' });
+        return;
+    }
     try {
         // Xóa Order_Confirmation nếu có
         await prisma.order_Confirmation.deleteMany({
